add a bunch of new functions to libt core and improve module system
The module system now allows loading modules with version restrictions,
and expects that all modules export a VERSION table. This table is
special and must be constructed with `t.Version.new` as it sets a
metatable for lexicographically ordering the versions.

Versions should use semantic versioning represented as a Lua table:

    module.VERSION = { major, minor, patch, revision }

The module cache can have multiple versions of the same library cached
at once. I'm not sure if this capability will ever get used or if it is
useful, but it's there.

New libt core functions:

 * find
 * packC, unpackC
 * any, all
 * id
 * fold
 * min
 * len
 * zip, izip
 * call, apply
 * eq, isNil