fix module caching, versioning, and some functions
Module caching was essentially broken. Now it actually works, yay ! We
avoid the problem of using tables as indices by instead serializing the
version tables before using them as keys, and deserializing them if we
need the key when we iterate.

The zip and min functions had some serious flaws which are now fixed,
too.