add libturtle
The goal of libturtle is to provide a unified interface for turtle functions.
Thanks to libturtle, it is possible to more easily iterate over directions and
perform actions on all those directions. For example, we can dig upwards, forwards, and downwards in a simple look like this:

  local t = loadfile("libt")
  local dir = t.requireStrict("libdirection")
  local turtle = t.requireStrict("libturtle")

  for d in t.from({dir.UP, dir.FORWARD, dir.DOWN}) do
    turtle.dig(d)
  end

The following turtle functions are overridden to accept directions now:

 * dig
 * inspect
 * place
 * compare
 * drop
 * suck
 * detect
 * attack

The following new turtle functions are added:

 * move (replaces up, down, forward, back)
 * turn (replaces turnLeft, turnRight)
   using the pseudo-direction dir.AROUND, turn will make the turtle turn right
   twice.

libturtle's metatable is such that requesting any function not present in
libturtle proper will perform a lookup in ComputerCraft's builtin turtle
module.