libtmsg: add optional JSON support
Tries to load `libjson`. If successful, defines sendJSON and recvJSON that
encode and decode their input and output respectively to and from JSON. These
methods accordingly use the application/json content-type in the request
header. The `sendJSON` function also has a parameter `parse` to set the accept
header to `application/json` as well. This setting is set to false by default.
In order to play nice with Servant routes that return JSON, it needs to be set
to true!