add rmid and two scripts to interact with it
Rmid is a remote code execution daemon. It listens on `/msg/X` for Lua tables
of the form `{ code = CODE, args = ARGS }` where CODE is a Lua program as a
string and ARGS are the arguments to supply to it. The CODE string is
frequently called the "program template" and ARGS the "template arguments". The
whole table is called the RMI table.

Simply put, rmid uses `loadstring` to convert the program template into a Lua
function, and invokes it with its template arguments. A machine running rmid is
said to be "rmid-compatible" or an "rmid server".

Using rmid, arbitrary code may be executed remotely, offering a lot of
flexibility in what can be achieved with libtmsg.

Two scripts that use rmid are included:

1. Enginectl sends a program template that will set the redstone output on a
given side to a given state. The template is sent to machines listening on
`/msg/<PREFIX><i>` where PREFIX is a customizable string (default is "ee") and
<i> ranges from i to COUNT. The purpose of enginectl is to control whether
electrical engines are to be active, but it could be used to control any
sequentially organized redstone-activated machines, such as lava pumps.

2. Eumon is designed to communicate with rmid servers with EU stores beneath
them. It uses the program template "reportpower", which gets the EU stored and
EU capacity of the attached device. This data is made into a table, and sent
using libtmsg to a given reply address. Eumon listens on this reply address
after sending the RMI table. It sums the tables componentwise to compute the
total EU stored and total EU capacity of the target devices. This information
is then printed and displayed on an attached monitor.