add libreply and librmi
Libreply contains a utility for manipulating code strings in order to create
a general request-reponse layer on top libtmsg in conjunction with rmid.

Librmi establishes a convention of storing "code templates" under the "rmi"
directory. Code templates are loaded automatically and converted into
request-response code using libreply. Librmi also caches code templates to
avoid disk IO on repeated calls.

Librmi's core function is `request` which sends a code template to a given
endpoint with some arguments. The result is as if pcall had been used to call
the function resulting from `loadfile` on the code template with those
arguments on the remote machine.

Librmi provides one more layer of abstraction on top of this via the `bind`
function. This function takes one argument that is a libtmsg endpoint, and
creates a proxy object for an rmid-compatible machine listening there. Function
calls on this object will invoke `request` using the name of the function as
the name of the code template to use. The arguments of the function are used as
the arguments of the code template. Requests made via a proxy object
transparently handle exceptions; if the result of the inner call to `request`
indicates failure, then the exception is re-raised on the client machine.