make libtput, tg, and tp "freestanding"
libtput is now made to only optionally depend on libt (for the list
and ilist functions), so it is possible to just directly loadfile
libtput from tg and tp. This makes it possible to use tg to download
libt, so the only things that need to be uploaded to pastebin to
bootstrap a libt installation is lib/libtput and tg.